requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
dj-rest-auth==6.0.0
Django>=4.2,<5
django-allauth==0.63.6
django-apscheduler==0.6.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysqlclient==2.2.4
openai==1.37.1
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
#PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
whitenoise==6.7.0
mysqlclient==2.2.4
openai==1.37.1
